Main Menu Mock Up and Proof of Concept loading screen.
Created a mock up for the main menu. Hovering over the buttons enables little running man next to the button. Clicking on the button causes the button to fly off screen essentially acting as a transition to the next screen.
The loading bar currently works based off an animation as there isn't anything to load in. I think it would be cool if there was concept art on the loading screen.