Quest Rework Update #1
Created a new object called QuestDisplayData.
Renamed panels in the message/quest app.
Reached the point where you can interact with an NPC, receive a quest in the app, click on the app which takes you the the QuestInfoPanel that displays the quest's information.
Next steps is to implement accepting quests and returning back to the quest list panel from the quest info panel.