I did stuff (ui related)
Got the mask to work so that the phone's background sits inside the phone's frame rather than resting on top. From what I understand about masks, everything that is transparent is hidden by the mask component, there in the screen gameobject I set the image to be a white png in the shape of the frame's interior.
Also, added in the 3 buttons at the bottom of the phone. Currently they all have the same functionality (go back to the home screen).
The tip app is complete with the new ui designed by Kyra. I made the lightbulb icon interactable so that players can press the light bulb to get a new tip rather than exiting and opening the app repeatedly.
The search bar in the homescreen now displays the name of the app your mouse hovers over.