Main Menu UI Updates
Fixed the scaling with the buttons on the main menu.
Swapped Bea and Nyx's profile picture.
Changed the UI to make Bea the default selected character.
Fixed scaling on pfp and confirm button in the character select screen.
Added the character selection overlay to the character selection screen.
Stuff I still need to do:
Set Bea as the default selected character in the code.
Insert the 3d models in the ui and have them change accordingly.