UI Update #2
Moved the quit race button to the quit panel. The quit panel wasn't being used, so I brought it back for the quit race button.
When the player opens their phone during a race, they will be show the quit race button. Pressing the back or home button will close the phone.